rename test_set module to normal_test
